+++ packages/search-ui-elasticsearch-connector/src/queryBuilders/ResultsAutocompleteBuilder.ts
@@ -3,7 +3,7 @@ import type {
   ResultSuggestionConfiguration
 } from "@elastic/search-ui";
 import { BaseQueryBuilder } from "./BaseQueryBuilder";
-import { getQueryFields } from "../ElasticsearchQueryTransformer/utils";
+import { getQueryFields } from "../utils";
 import { SearchRequest } from "../types";
 
 export class ResultsAutocompleteBuilder extends BaseQueryBuilder {
